+++ model/facility.go
@@ -2,11 +2,11 @@ package model
 
 type Facility struct {
 	Recipe   *Recipe
-	Amount   uint8
+	Amount   float64
 	Children *[]FacilityChild
 }
 
 type FacilityChild struct {
-	Remain   uint8
+	Remain   float64
 	Facility *Facility
 }
